Move is_rust_identifier to utils module
