+++ src/type_id.rs
@@ -14,7 +14,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::form::{Form, FreeForm};
+use crate::{
+	utils::is_rust_identifier,
+	form::{
+		Form,
+		FreeForm,
+	},
+};
 use derive_more::From;
 use serde::Serialize;
 
@@ -59,26 +65,6 @@ impl Namespace {
 			return Err(NamespaceError::MissingSegments);
 		}
 		if let Some(err_at) = segments.iter().position(|seg| {
-			/// Returns `true` if the given string is a proper Rust identifier.
-			fn is_rust_identifier(s: &str) -> bool {
-				// Only ascii encoding is allowed.
-				// Note: Maybe this check is superseeded by the `head` and `tail` check.
-				println!("is_rust_identifier? \"{}\"", s);
-				if !s.is_ascii() {
-					return false;
-				}
-				if let Some((&head, tail)) = s.as_bytes().split_first() {
-					// Check if head and tail make up a proper Rust identifier.
-					let head_ok = head == b'_' || head >= b'a' && head <= b'z' || head >= b'A' && head <= b'Z';
-					let tail_ok = tail.iter().all(|&ch| {
-						ch == b'_' || ch >= b'a' && ch <= b'z' || ch >= b'A' && ch <= b'Z' || ch >= b'0' && ch <= b'9'
-					});
-					head_ok && tail_ok
-				} else {
-					// String is empty and thus not a valid Rust identifier.
-					false
-				}
-			}
 			!is_rust_identifier(seg)
 		}) {
 			return Err(NamespaceError::InvalidIdentifier { segment: err_at });
